fixed package-lock.json, moved login into its own dir
