+++ school-cal/src/AppRouter.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-
+import Login from "./components/Login";
 const AppRouter = () => {
   return (
     <>
       <Router>
         <Switch>
           <Route exact path="/" /> {/* For a future landing page */}
-          <Route path="/login" />
+          <Route path="/login" component={Login} />
           <Route path="/register" />
         </Switch>
       </Router>
