+++ school-cal/src/firebase/index.js
@@ -3,16 +3,17 @@ import "firebase/auth";
 import "firebase/firestore";
 import "firebase/functions";
 
+console.log(process.env.CREATE_REACT_API_KEY);
 export const config = {
-    apiKey: "AIzaSyCKHnk_AIoXYU7BcUa63LhQQWa6-Ma3ujQ",
-    authDomain: "school-calendar-staging.firebaseapp.com",
-    databaseURL: "https://school-calendar-staging.firebaseio.com",
-    projectId: "school-calendar-staging",
-    storageBucket: "school-calendar-staging.appspot.com",
-    messagingSenderId: "482714059970",
-    appId: "1:482714059970:web:b34c56dcb33f214576675b",
-    measurementId:"G-2NCWESMGBP"
-  };
+  apiKey: process.env.REACT_APP_API_KEY,
+  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
+  databaseURL: process.env.REACT_APP_DATABASE_URL,
+  projectId: process.env.REACT_APP_PROJECT_ID,
+  storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_APP_ID,
+  measurementId: process.env.REACT_APP_MEASUREMENT_ID
+};
 
 export const app = firebase.initializeApp(config);
 export const db = firebase.firestore();
